Added a functionality to show CRUD only to logged users

+++ app/views/director.view.php
@@ -11,8 +11,11 @@
             foreach ($directores as $director) {
                 echo "<li>";
                 echo $director->nombre;
-                echo "<a href='verPeliculasDirector/" . $director->id . "' class='a-animation'>Ver películas</a>";
-                echo "<a href='eliminarDirector/" . $director->id . "' class='a-animation'>Borrar</a>";
+                if (isset($_SESSION['USER'])) {
+                    echo "<a href='verPeliculasDirector/" . $director->id . "' class='a-animation'>Ver películas</a>";
+                    echo "<a href='eliminarDirector/" . $director->id . "' class='a-animation'>Borrar</a>";
+                }
+
                 echo "<img src='{$director->imagen}' width='200'>";
                 echo "</li>";
             }
